+++ js/controllers/levels.js
@@ -5,8 +5,6 @@ import { URL } from '../utils/url';
 const handlePrograms = require('./programs');
 
 const handleLevels = (userId, token) => {
-  console.log('user', userId);
-
   $.ajax({
     url: `${URL}/levels/${userId}`,
     type: 'GET',
@@ -16,9 +14,9 @@ const handleLevels = (userId, token) => {
     success: function (data) {
       console.log('LEVELS', data);
 
-      const { name, description } = data;
+      const { programId, name, description } = data;
 
-      handlePrograms(data.programId, token, name, description)
+      handlePrograms(programId, token, name, description)
     }
   });
 };
