+++ scss/style.scss
@@ -11,3 +11,13 @@
 
 // utils
 @import './utils/load';
+
+.box {
+  background-color: $white;
+  border-radius: 8px;
+  height: 386px;
+  overflow-y: scroll;
+  width: 300px;
+  margin: ($gutter * 3) ($gutter * 3) 0 0;
+  padding: $gutter;
+}
